Fixed smart pointer use

+++ src/lib/gui/QDataStreamProxy.h
@@ -28,7 +28,7 @@ public:
   }
   virtual ~QDataStreamProxy() = default;
 
-  virtual qint64 writeRawData(const char *data, qint64 len) {
+  virtual qint64 writeRawData(const char *data, int len) {
     assert(m_Stream);
     return m_Stream->writeRawData(data, len);
   }
